requirements.txt
asgiref==3.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
Pillow==9.1.0
sqlparse==0.4.2
tzdata==2022.1
